requirements.txt
# Core dependencies
PyQt6>=6.4.0
paramiko>=3.4.0
pycryptodome>=3.20.0
psutil>=5.9.0

# Testing
pytest>=8.0.0
pytest-qt>=4.2.0
pytest-cov>=4.1.0

python-dotenv==1.0.1
typing-extensions==4.9.0
black==24.1.1
pylint==3.0.3
mypy>=1.8.0 